+++ src/Form/order/OrderItemType.php
@@ -1,27 +1,27 @@
 <?php
 
-namespace App\Form;
+namespace App\Form\order;
 
-use App\Entity\CreditCardDetails;
+use App\Entity\OrderItem;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
-class CreditCardDetailsType extends AbstractType
+class OrderItemType extends AbstractType
 {
     public function buildForm(FormBuilderInterface $builder, array $options)
     {
         $builder
-            ->add('code')
-            ->add('cvv')
-            ->add('expiresAt')
+            ->add('productCode')
+            ->add('amount')
+            ->add('price')
         ;
     }
 
     public function configureOptions(OptionsResolver $resolver)
     {
         $resolver->setDefaults([
-            'data_class' => CreditCardDetails::class,
+            'data_class' => OrderItem::class,
         ]);
     }
 }
